[BUGFIX] Mutex redis behavior related to `redsync` pkg
Adding configurations for Expiration and Retries

Sanitizing  `redsync` defaults in the Redis Mutex factory. The library sets custom defaults that might cause unexpected behaviors. Lock timeout, for instance, is based in expiration setting and set to 8 seconds, waiting that time until the lock expires in redis instead of returning error for not being able to acquire the lock, btw does 32 retries by default. This means that if the mutex is used with these defaults on a heavy load task in would be breaking the mutual exclusion if it takes more than 8 seconds

However, new configuration parameters allow to still keep that behavior if needed.

```
// NewMutex returns a new distributed mutex with given name.
func (r *Redsync) NewMutex(name string, options ...Option) *Mutex {
	m := &Mutex{
		name:   name,
		expiry: 8 * time.Second,
		tries:  32,
		delayFunc: func(tries int) time.Duration {
			return time.Duration(rand.Intn(maxRetryDelayMilliSec-minRetryDelayMilliSec)+minRetryDelayMilliSec) * time.Millisecond
		},
		genValueFunc:  genValue,
		driftFactor:   0.01,
		timeoutFactor: 0.05,
		quorum:        len(r.pools)/2 + 1,
		pools:         r.pools,
	}
	...
```